requirements.txt
# Core Dependencies
torch>=2.0.0
torchtext>=0.15.0
transformers>=4.30.0
tokenizers>=0.13.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
sacrebleu>=2.3.0
sentencepiece>=0.1.99

# NLP Tools
spacy>=3.5.0
jieba>=0.42.1
sacremoses>=0.0.53

# Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
einops>=0.6.1

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: Experiment Tracking
# wandb>=0.15.0
# tensorboard>=2.13.0

